+++ src/components/Card.tsx
@@ -1,11 +1,8 @@
 import styled from 'styled-components'
 
-import { ReactComponent as Heart } from '../assets/icons/heart.svg'
-import { ReactComponent as Message } from '../assets/icons/message-circle.svg'
-import { ReactComponent as Cart } from '../assets/icons/shopping-cart.svg'
 import { ColorIcon } from '../elements/ColorIcons'
 import { theme } from '../style/theme'
-import { DataType } from '../pages/ListPage'
+import { DataType } from '../containers/listPage/CardList'
 
 const Card = (props: DataType) => {
   return (
@@ -20,32 +17,6 @@ const Card = (props: DataType) => {
             [{props.p_Brand}]{props.p_Name}
           </CardTitle>
           <Cardprice>1,900원</Cardprice>
-        </Div>
-        <Div>
-          <IconDiv>
-            <IconSpan>
-              <Message
-                style={{
-                  color: `${theme.colors.gray2}`,
-                }}
-              />
-              <Count>{props.p_Review}</Count>
-            </IconSpan>
-            <IconSpan>
-              <Heart
-                style={{
-                  color: `${theme.colors.gray2}`,
-                  width: '20px',
-                }}
-              />
-              <Count>{props.p_Like}</Count>
-            </IconSpan>
-            <Cart
-              style={{
-                color: `${theme.colors.gray2}`,
-              }}
-            />
-          </IconDiv>
         </Div>
       </CardCp>
     </>
@@ -86,28 +57,4 @@ const CardTitle = styled.div`
 const Cardprice = styled.div`
   font-size: ${theme.fontSize.small};
   color: ${theme.colors.green2};
-`
-const Count = styled.span`
-  font-size: ${theme.fontSize.small};
-  color: ${theme.colors.gray2};
-`
-const IconSpan = styled.span`
-  cursor: pointer;
-  display: inline-flex !important;
-  align-items: center;
-  justify-content: flex-end;
-  flex-direction: row;
-  gap: 5px;
-  margin: 0 !important;
-`
-const IconDiv = styled.div`
-  display: inline-flex;
-  gap: 10px;
-  margin-top: 0.75em !important;
-  flex-direction: row;
-  justify-content: space-evenly;
-  align-content: center;
-  flex-wrap: wrap;
-  align-items: center;
-  text-align: center;
 `
